delete RedshiftDestination and RedshiftInsertDestination
